Move to OCSController
