+++ appinfo/application.php
@@ -35,37 +35,5 @@ class Application extends \OCP\AppFramework\App {
 	 */
 	public function __construct($urlParams = array()) {
 		parent::__construct('survey_server', $urlParams);
-		$this->registerService();
 	}
-
-	private function registerService() {
-		$container = $this->getContainer();
-
-		$container->registerService('statisticService', function() {
-			return new StatisticService(\OC::$server->getDatabaseConnection(), \OC::$server->getConfig());
-		});
-	}
-
-	/**
-	 * register OCS API Calls
-	 */
-	public function registerOCSApi() {
-
-		$container = $this->getContainer();
-		$server = $container->getServer();
-
-		$request = $server->getRequest();
-		$statisticService = $container->query('statisticService');
-		$api = new ExternalApi($request, $statisticService, $server->getConfig());
-		//$api = new ExternalApi($server->getRequest(), $container->query('statisticService'));
-
-		API::register('post',
-			'/apps/survey_server/api/v1/survey',
-			array($api, 'receiveSurveyResults'),
-			'survey_server',
-			API::GUEST_AUTH
-		);
-
-	}
-
 }
