+++ appinfo/routes.php
@@ -28,12 +28,15 @@
  * it's instantiated in there
  */
 return [
-    'routes' => [
-	   ['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
-	   ['name' => 'page#do_echo', 'url' => '/echo', 'verb' => 'POST'],
+	'routes' => [
+		['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
+		['name' => 'page#do_echo', 'url' => '/echo', 'verb' => 'POST'],
 
 		// api
 		['name' => 'api#add', 'url' => '/api/v1/data', 'verb' => 'POST'],
 		['name' => 'api#get', 'url' => '/api/v1/data', 'verb' => 'GET'],
-    ]
+	],
+	'ocs' => [
+		['name' => 'ExternalApi#receiveSurveyResults', 'url' => '/api/v1/survey', 'verb' => 'POST'],
+	]
 ];
